+++ app/src/main/java/woowacourse/shopping/data/product/dataSource/LocalProductsDataSource.kt
@@ -1,9 +1,10 @@
-package woowacourse.shopping.view.data
+package woowacourse.shopping.data.product.dataSource
 
 import android.content.Context
 import androidx.room.Room
+import woowacourse.shopping.data.product.dao.ProductDao
+import woowacourse.shopping.data.product.database.ProductDatabase
 import woowacourse.shopping.data.product.entity.ProductEntity
-import woowacourse.shopping.data.product.storage.ProductsDataSource
 
 object LocalProductsDataSource : ProductsDataSource {
     private lateinit var dao: ProductDao
