+++ app/src/main/java/woowacourse/shopping/data/product/database/ProductDatabase.kt
@@ -1,7 +1,8 @@
-package woowacourse.shopping.view.data
+package woowacourse.shopping.data.product.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import woowacourse.shopping.data.product.dao.ProductDao
 import woowacourse.shopping.data.product.entity.ProductEntity
 
 @Database(entities = [ProductEntity::class], version = 1)
